Type: Feature
Brief: Class Loggable implemented

Detailed:
I have created a class "Loggable" as a wrapper to winston, enabling any child class of him to "log" into the console and a file.
The Loggable class can be configured on .env (as shown on the .env comments). Timezone supported on the console, while file log prints UTC timestamps.